Use correct paths to find configs; allow aliased generator names
To access the config files in the directory where the user is calling
the command, we needed to use 'process.cwd()' rather than just using a
relative path.

The first generator I'm writing is for react components. After writing
> accio react_component component_name
several times, I feel like having the shortcut 'rc' for
'react_component', so I added the option to *alias* the generator name
in the configuration file.